requirements.txt
flask>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-groq>=0.1.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
python-dotenv>=1.0.0
groq>=0.4.0
werkzeug>=3.0.0
pdf2image>=1.17.0
flask-cors>=4.0.0
